Moving python3 code in separate files to try to fix tests for python2
