+++ gql/__init__.py
@@ -1,4 +1,12 @@
-from .gql import gql
-from .client import Client, AsyncClient
+import sys
 
-__all__ = ["gql", "Client", "AsyncClient"]
+from .gql import gql
+from .client import Client
+
+__all__ = ["gql", "Client"]
+
+if sys.version_info > (3, 6):
+    from .async_client import AsyncClient
+
+    # Cannot use __all__.append here because of flake8 warning
+    __all__ = ["gql", "Client", "AsyncClient"]
