+++ gql/transport/__init__.py
@@ -1,73 +1,12 @@
-import abc
-from typing import Union
+import sys
 
-import six
-from graphql.execution import ExecutionResult
-from graphql.language.ast import Document
-from promise import Promise
+from .transport import Transport
 
-from typing import Dict, Optional, AsyncGenerator
+__all__ = ["Transport"]
 
 
-@six.add_metaclass(abc.ABCMeta)
-class Transport:
-    @abc.abstractmethod
-    def execute(self, document):
-        # type: (Document) -> Union[ExecutionResult, Promise[ExecutionResult]]
-        """Execute the provided document AST for either a remote or local GraphQL Schema.
+if sys.version_info > (3, 6):
+    from .async_transport import AsyncTransport
 
-        :param document: GraphQL query as AST Node or Document object.
-        :return: Either ExecutionResult or a Promise that resolves to ExecutionResult object.
-        """
-        raise NotImplementedError(
-            "Any Transport subclass must implement execute method"
-        )
-
-
-@six.add_metaclass(abc.ABCMeta)
-class AsyncTransport:
-    @abc.abstractmethod
-    async def connect(self):
-        """Coroutine used to create a connection to the specified address
-        """
-        raise NotImplementedError(
-            "Any AsyncTransport subclass must implement execute method"
-        )
-
-    @abc.abstractmethod
-    async def close(self):
-        """Coroutine used to Close an established connection
-        """
-        raise NotImplementedError(
-            "Any AsyncTransport subclass must implement execute method"
-        )
-
-    @abc.abstractmethod
-    async def execute(
-        self,
-        document: Document,
-        variable_values: Optional[Dict[str, str]] = None,
-        operation_name: Optional[str] = None,
-    ) -> ExecutionResult:
-        """Execute the provided document AST for either a remote or local GraphQL Schema.
-        """
-        raise NotImplementedError(
-            "Any AsyncTransport subclass must implement execute method"
-        )
-
-    @abc.abstractmethod
-    def subscribe(
-        self,
-        document: Document,
-        variable_values: Optional[Dict[str, str]] = None,
-        operation_name: Optional[str] = None,
-    ) -> AsyncGenerator[ExecutionResult, None]:
-        """Send a query and receive the results using an async generator
-
-        The query can be a graphql query, mutation or subscription
-
-        The results are sent as an ExecutionResult object
-        """
-        raise NotImplementedError(
-            "Any AsyncTransport subclass must implement execute method"
-        )
+    # Cannot use __all__.append here because of flake8 warning
+    __all__ = ["Transport", "AsyncTransport"]
